Implement group selection in multiselects.
This is needed as a basis for #68, #78, #115, #124.
This change allows us to simply specify group options in the multiselect
dropdowns as such:

@component('components.form.multiSelectInput', [
  'name' => 'participants',
  // ...
  'options' => $course->participants->all(),
  'groups' => ['Select all' => $course->participants->all(), 'First and last one' => [$course->participants->first(), $course->participants->last()]],
  'valueFn' => // ...
])@endcomponent

The groups will be rendered in the order they are given after all normal
options (search is possible of course). When the user clicks on a group,
all previously selected options are replaced with the options from the
group.

Empty groups and groups in single-selects (aka aliases) are also
supported.