Adapt tests to new form value restoration flow
The form restoration is now driven by a TokenMismatchException, since in
Laravel 7 the CSRF protection middleware runs before the authentication
check middleware. Therefore we need to mock that middleware in the test.

Also, we now redirect back instead of to the login page after saving
the form contents, in order to allow saved forms even when not logged in
(e.g. saving form data on the sentry error form).