Accept more import formats, and report errors correctly
We had a sentry issue where someone apparently renamed an xlsx file to
xls and tried to upload it in the block import form.

This commit delegates to PhpSpreadsheet the task of finding the right
Reader class (and in the case of CSV the right encoding) to parse an
uploaded spreadsheet. This is now done the same way for both
participants and blocks import. Also, the error reporting is more
consistent between the two now.