Rename qualis to feedbacks
I also had to monkey-patch the InflectorFactory which is responsible for
auto-converting between singular and plural names in the code. Without
this change, the FeedbackDataFactory would look for the list of its
associated feedbacks under the name FeedbackData#feedback, because in
normal english, feedback has no plural. However, we are really talking
about multiple separate feedback conversations, so in the world of
Qualix, feedbacks is actually the correct term to use in the code.