Sync changes from the quali editor to the requirements matrix
This is done using an invisible editor which includes the tiptap / yjs
syncing capabilities. This way, a copy of the feedback contents is kept
up to date for each row in the matrix, and that copy influences the
requirement statuses which are displayed.

Syncing in the other direction (requirements matrix to quali editors)
will follow in a separate commit.